+++ backend/src/main/java/com/hellweek/coffee/dto/analytics/SalesAnalytics.java
@@ -3,12 +3,13 @@ package com.hellweek.coffee.dto.analytics;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class SalesAnalytics {
-    private double totalRevenue;
+    private BigDecimal totalRevenue;
     private int totalOrders;
-    private double averageOrderValue;
+    private BigDecimal averageOrderValue;
 }
